101/integration/requirements.txt
Flask==2.3.3
gunicorn==20.1.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
mysql-connector-python
